chore(weave): extract clickhouse/annotation_queues.py with CHInfraProtocol
Move all annotation queue CRUD methods into AnnotationQueuesMixin.
Add CHInfraProtocol for mixin type safety.
